Start for Combat System
Start for base implementation of combat system.

Attacks will send over a form of attack data as well as who the attack originated from.

Lots of low level designing so that system can be expanded on. Most work is intentionally made very abstract so that it can be made into whatever as development continues